refactor(popup): App.svelte router and typescript
