+++ package.json
@@ -5,22 +5,18 @@
   "devDependencies": {
     "@catppuccin/tailwindcss": "^0.1.6",
     "@sveltejs/vite-plugin-svelte": "^3.1.1",
-    "@types/webextension-polyfill": "^0.10.7",
     "autoprefixer": "^10.4.19",
     "lucide-svelte": "^0.383.0",
     "postcss": "^8.4.38",
-    "prettier": "^3.3.0",
-    "prettier-plugin-svelte": "^3.2.3",
+    "prettier": "^3.3.1",
+    "prettier-plugin-svelte": "^3.2.4",
     "prettier-plugin-tailwindcss": "^0.5.14",
-    "rollup-plugin-copy": "^3.5.0",
-    "svelte": "^4.2.17",
+    "svelte": "^4.2.18",
     "svelte-check": "^3.8.0",
-    "svelte-routing": "^2.13.0",
+    "svelte-spa-router": "^4.0.1",
     "tailwindcss": "^3.4.4",
-    "tinro": "^0.6.12",
     "tslib": "^2.6.3",
     "typescript": "^5.4.5",
-    "webextension-polyfill": "^0.12.0",
     "wxt": "^0.18.4"
   },
   "description": "Schooltape is a free, open-source, web extension that allows you to customise the look, feel, and functionality of Schoolbox!",
@@ -35,8 +31,7 @@
     "zip": "wxt zip",
     "zip:firefox": "wxt zip -b firefox",
     "check": "svelte-check --tsconfig ./tsconfig.json",
-    "format": "prettier --write .",
-    "postinstall": "wxt prepare"
+    "format": "prettier --write ."
   },
   "type": "module"
 }
