+++ src/entrypoints/popup/routes/Home.svelte
