+++ src/entrypoints/popup/routes/Plugins.svelte
