+++ src/entrypoints/popup/routes/Snippets.svelte
